penrose-app/backend/requirements.txt
Django>=4.2
openai>=1.40.0
Pillow>=10.3.0
